Relacionamentos criados

+++ src/main/java/br/com/carro/repositories/CarroRepository.java
@@ -3,5 +3,5 @@ package br.com.carro.repositories;
 import br.com.carro.entities.Carro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarroRepository extends JpaRepository<Carro,Integer> {
+public interface CarroRepository extends JpaRepository<Carro,Long> {
 }
